requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gitpython==3.1.41
groq==0.5.0
python-dotenv==1.0.0
langgraph==0.2.0
